requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-jazzmin==2.4.8
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
